# cmd/api/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  models.ProcessInfo:
    properties:
      command:
        type: string
      cpu:
        type: number
      memory:
        type: integer
      name:
        type: string
      pid:
        type: integer
      state:
        type: string
    type: object
...
  ssl.CreateSSLRequest:
    properties:
      acme_client_id:
        type: integer
      domain:
        type: string
    required:
    - acme_client_id
    - domain
    type: object
host: localhost:8080
info:
  contact:
    email: support@etapanel.com
    name: EtaPanel Support
    url: https://github.com/EtaPanel-dev/EtaPanel
  description: EtaPanel 服务器管理面板后端API文档，提供系统监控、文件管理、Nginx配置、SSL证书管理等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EtaPanel API
  version: "1.0"
paths:
...
  /auth/acme/ssl:
    get:
      consumes:
      - application/json
      description: 获取所有SSL证书信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 获取SSL证书列表
      tags:
      - SSL证书管理
    post:
      consumes:
      - application/json
      description: 通过ACME协议申请SSL证书
      parameters:
      - description: SSL申请信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ssl.CreateSSLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 申请成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 申请SSL证书
      tags:
      - SSL证书管理
  /auth/acme/ssl/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的SSL证书
      parameters:
      - description: 证书ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 证书不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 删除SSL证书
      tags:
      - SSL证书管理
    put:
      consumes:
      - application/json
      description: 更新指定ID的SSL证书信息
      parameters:
      - description: 证书ID
        in: path
        name: id
        required: true
        type: integer
      - description: SSL证书信息
        in: body
        name: request
        required: true
        schema:
          properties:
            auto_renew:
              type: boolean
            certificate:
              type: string
            domain:
              type: string
            email:
              type: string
            enabled:
              type: boolean
            key:
              type: string
            provider:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 证书不存在
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 更新SSL证书
      tags:
      - SSL证书管理
...
  /auth/system/process/kill:
    post:
      consumes:
      - application/json
      description: 向指定进程发送信号以终止进程
      parameters:
      - description: 进程ID和信号类型
        in: body
        name: request
        required: true
        schema:
          properties:
            pid:
              type: integer
            signal:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 终止成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "403":
          description: 无法终止系统关键进程
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 终止进程
      tags:
      - 系统监控
  /auth/system/processes:
    get:
      consumes:
      - application/json
      description: 获取系统中所有运行的进程信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  properties:
                    processes:
                      items:
                        $ref: '#/definitions/models.ProcessInfo'
                      type: array
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: 获取进程列表
      tags:
      - 系统监控
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入Bearer token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...